+++ llama/src/main/cpp/llama-android.cpp
@@ -312,7 +312,7 @@ Java_android_llama_cpp_LLamaAndroid_free_1batch(JNIEnv *, jobject, jlong batch_p
 
 extern "C"
 JNIEXPORT jlong JNICALL
-Java_android_llama_cpp_LLamaAndroid_new_1sampler(JNIEnv *, jobject,jint tempe) {
+Java_android_llama_cpp_LLamaAndroid_new_1sampler(JNIEnv *, jobject,jfloat tempe) {
     auto sparams = llama_sampler_chain_default_params();
     sparams.no_perf = true;
     llama_sampler * smpl = llama_sampler_chain_init(sparams);
